♻️ Refactor(battle): restructure battle tab components; create ActionSection, CharacterBattleTabContent, PlayerBattleTabContent, and NPCBattleTabContent for improved character combat display
